requirements.txt
ptrx-config==7.0.5

aiofiles==0.7.0
click==7.1.2
fastapi==0.65.1
gunicorn==20.1.0
h11==0.12.0
pydantic==1.8.2
PyYAML==5.4.1
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.13.4